Rename datetime javascript helper
